refactor: move the places handler files into the Wayfinding folder, where they are used.
