+++ packages/map-template/src/components/Wayfinding/Wayfinding.jsx
@@ -17,7 +17,7 @@ import SearchField from '../WebComponentWrappers/Search/Search';
 import { snapPoints } from '../../constants/snapPoints';
 import { usePreventSwipe } from '../../hooks/usePreventSwipe';
 import generateMyPositionLocation from '../../helpers/MyPositionLocation';
-import addGooglePlaceGeometry from "../MapWrapper/GoogleMapsMap/GooglePlacesHandler";
+import addGooglePlaceGeometry from "./googlePlacesHandler";
 import GooglePlaces from '../../assets/google-places.png';
 import { mapTypes } from "../../constants/mapTypes";
 import { ReactComponent as WalkIcon } from '../../assets/walk.svg';
@@ -27,7 +27,7 @@ import { ReactComponent as CompassArrow } from '../../assets/compass-arrow.svg';
 import { travelModes } from "../../constants/travelModes";
 import Dropdown from "../WebComponentWrappers/Dropdown/Dropdown";
 import primaryColorState from "../../atoms/primaryColorState";
-import addMapboxPlaceGeometry from "../MapWrapper/MapboxMap/MapboxPlacesHandler";
+import addMapboxPlaceGeometry from "./mapboxPlacesHandler";
 import mapboxAccessTokenState from "../../atoms/mapboxAccessTokenState";
 import distanceUnitSystemSelector from '../../selectors/distanceUnitSystemSelector';
 import useDirectionsInfo from "../../hooks/useDirectionsInfo";
