+++ packages/map-template/src/components/Wayfinding/googlePlacesHandler.js
