+++ packages/map-template/src/components/Wayfinding/mapboxPlacesHandler.js
